requirements.txt
torch
torchvision
pandas
numpy
matplotlib
Pillow
opencv-python
tifffile
imagecodecs